+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Add WildStacker integration
  - Add "biomes" and "not_biomes" options to spawn rules
+ - Add support for "" or "*" target entity types (wildcard, will match all entity types)
 
 # 1.7